Work with code_14

+++ t_12/TouchGFX/gui/include/gui/graphic_1__screen/Graphic_1_Presenter.hpp
@@ -28,7 +28,7 @@ public:
 
 
     virtual void valueIsChanged();
-    void userToModel(uint8_t value);
+    void userToModel(uint16_t value);
     void modelToView();
